fix: 게시글 조회 쿼리 N+1 문제 수정 및 최적화

+++ server/src/main/java/teamdropin/server/domain/box/dto/box/BoxCreateRequestDto.java
@@ -26,10 +26,10 @@ public class BoxCreateRequestDto {
     private String phoneNumber;
 
     @NotNull
-    private Integer cost;
+    private Long cost;
 
     @NotNull
-    private Integer area;
+    private Long area;
 
     private boolean barbellDrop;
     private String url;
