+++ server/src/main/java/teamdropin/server/domain/box/dto/box/UpdateBoxRequestDto.java
@@ -27,10 +27,10 @@ public class UpdateBoxRequestDto {
     private String phoneNumber;
 
     @NotNull
-    private Integer cost;
+    private Long cost;
 
     @NotNull
-    private Integer area;
+    private Long area;
 
     private boolean barbellDrop;
     private String url;
